Added logging, updated project and related configurations
Ignore bin dir.

This project now depends on the parallel current forked version of gradle-stdproject-plugin.
A bunch of stuff had to go into the buildscript classpath to make this work properly, repos are added to the settings.gradle pluginManagement block. gradle-stdproject-plugin should be in a directory adjacent to this when building so Gradle picks it up correctly.

Enable caching, set reasonable gradle daemon JVM args.

Added eclipse buildship support to the project, it now productes proper eclipse configurations with spotbugs.

Produces proper sources and javadoc artifacts now.

Add logging to help with figuring out the Velocity context settings. Was very helpful when initially setting up the plugin for me.